+++ projects/lnk-shared-ui/src/lib/directives/index.ts
@@ -17,3 +17,4 @@
 export * from './infinite-scroll.directive';
 export * from './list-sticky-toolbar.directive';
 export * from './required-label.directive';
+export * from './scrollable-region-focusable.directive';